添加baseDao

+++ src/main/java/com/zw/weixin/service/UserService.java
@@ -3,5 +3,5 @@ package com.zw.weixin.service;
 import com.zw.weixin.bean.User;
 
 public interface UserService {
-	public User getById(int id);
+	public User getById(Long id);
 }
